backend/foodgram_backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.5
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
urllib3==2.5.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.9
